% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{bar_plot_highcharts}
\alias{bar_plot_highcharts}
\title{Highchart Bar plot}
\usage{
bar_plot_highcharts(
  df,
  x_var,
  y_var,
  horizontal = FALSE,
  fill_var = NULL,
  position = "dodge",
  title = "",
  y_lim = NULL,
  y_breaks = NULL,
  x_breaks = NULL,
  proportion = FALSE,
  scale_percentage = TRUE,
  other_vars = NULL,
  x_lab = NULL,
  y_lab = NULL,
  arrange_by = NULL,
  arrange_desc = TRUE,
  arrange_by_fill = NULL,
  fill_var_order = NULL,
  color_x_value = NULL,
  bar_size = NULL,
  normalize_prop = TRUE,
  break_x_var_names = FALSE,
  plot_height = NULL,
  group_color = lifecycle::deprecated(),
  legend_title = NULL,
  facet_by = NULL,
  reversed_stacks = FALSE,
  group_padding = NULL,
  add_total = FALSE,
  total_var = "total",
  text_size = NULL,
  remove_value = NULL,
  horizontal_line = NULL,
  n_decimals = 0,
  x_var_order = NULL
)
}
\arguments{
\item{df}{data}

\item{x_var}{x-axis variable}

\item{y_var}{y_axis variable}

\item{horizontal}{whether or not the bar plot should be horizontal,}

\item{fill_var}{the name of the fill variable}

\item{position}{whether or not to make the bars stack or dodge according to
ggplot2 semantics}

\item{title}{title of the graph}

\item{y_lim}{limits of y-axis, if \code{NULL} and \code{proportion} is \code{TRUE}, will be
set to \code{c(0, 100)}}

\item{y_breaks}{breaks of y_axis}

\item{x_breaks}{breaks of x-axis}

\item{proportion}{if the values are percentages}

\item{scale_percentage}{if percentages should the scaled}

\item{other_vars}{other variables to include in the tooltip, should be a
named list where the name will be the key in the tooltip}

\item{x_lab}{labels on x axis}

\item{y_lab}{labels on y axis}

\item{arrange_by}{column to sort by}

\item{arrange_desc}{to arrange descending}

\item{arrange_by_fill}{what value in \code{fill_var} that sort be sorted by}

\item{fill_var_order}{what order \code{fill_var} should be displayed in, can
alternatively be \code{auto_character} or \code{auto_numeric} to automatically sort
the levels}

\item{color_x_value}{columns that should be different color, input as list.}

\item{bar_size}{width of bars}

\item{normalize_prop}{if _prop variable should be normalized, only valid
for stacked bar plot.}

\item{break_x_var_names}{if line breaks should be done for x_var}

\item{plot_height}{height of plot, value is in pixels}

\item{group_color}{deprecated}

\item{legend_title}{title of the legend}

\item{facet_by}{variable in \code{df} with at most 2 unique values to facet by;
if supplied the return value is a named list of plots, one per facet level}

\item{reversed_stacks}{should stacks be reversed?}

\item{group_padding}{distance between bars when using a dodge bar}

\item{add_total}{if total should be added to x-axis}

\item{total_var}{name of column that contains total}

\item{text_size}{size of text, will be interperted as pixels}

\item{remove_value}{value to remove, useful for removing obfuscated
observations}

\item{horizontal_line}{horizontal reference line}

\item{n_decimals}{number of decimals to round numbers to}

\item{x_var_order}{the order the x-variable should be displayed in}
}
\value{
highcharts config, or a named list of configs when \code{facet_by} is set
}
\description{
creates a bar plot in highchart, outputs the config file
}
