% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot.R
\name{box_plot}
\alias{box_plot}
\title{Produce a box plot}
\usage{
box_plot(
  df,
  x_var,
  y_var,
  fill_var = NULL,
  y_lower,
  y_upper,
  y_min,
  y_max,
  y_breaks = ggplot2::waiver(),
  y_labels = ggplot2::waiver(),
  y_lim = NULL,
  x_lab = NULL,
  y_lab = NULL,
  title = ggplot2::waiver(),
  horizontal = FALSE,
  arrange_by = x_var,
  arrange_desc = TRUE,
  add_total = FALSE,
  total_col = "total",
  fill_title = NULL,
  palette_type = "qualitative",
  colors = NULL,
  remove_grid = TRUE,
  text_size = 7
)
}
\arguments{
\item{df}{The data.frame to plot}

\item{x_var}{The name of the x variable}

\item{y_var}{The name of the median variable}

\item{fill_var}{The name of the fill variable}

\item{y_lower}{The name of the 25\% quantile variable}

\item{y_upper}{The name of the 75\% quantile variable}

\item{y_min}{The name of the 5\% quantile variable}

\item{y_max}{The name of the 95\% quantile variable}

\item{y_breaks}{The length of the skips between values in the y-axis}

\item{y_labels}{Whether or not to add labels to the plot}

\item{y_lim}{The limit for the y-axis}

\item{x_lab}{The label for the x-axis}

\item{y_lab}{The label for the y-axis}

\item{title}{The title of the graph}

\item{horizontal}{Whether the graph should be flipped}

\item{arrange_by}{The variable to order by}

\item{arrange_desc}{Whether to arrange in descending order}

\item{add_total}{If total should be added to labels on x-axis}

\item{total_col}{The column that should be used for adding total}

\item{fill_title}{The title of the fill legend}

\item{palette_type}{The type of palette to use, passed to
\code{\link[=colors_rc_2]{colors_rc_2()}}}

\item{colors}{manually specified colors to use for fill, only used when
\code{palette_type} is \code{qualitative}, must be subset of
\code{colors_rc_2(9, "qualitative")}}

\item{remove_grid}{if grid should be removed}

\item{text_size}{passed to \code{\link[=theme_rc]{theme_rc()}}}
}
\description{
Produces a box plot according to Registercentrum graphical standards
}
\examples{
df <- data.frame(
  y = c(rnorm(500, 50, 10), rnorm(500, 45, 10)),
  unit = rep(letters[1:2], each = 500)
) |>
  RCStat::get_aggregate_value(
    group_cols = "unit",
    vars = list(median = "y")
  )

box_plot(
  df = df,
  x_var = "unit",
  y_var = "y_median",
  y_lower = "y_quant_25",
  y_upper = "y_quant_75",
  y_min = "y_quant_5",
  y_max = "y_quant_95",
  arrange_by = "unit",
  arrange_desc = FALSE
)

box_plot(
  df = df,
  x_var = "unit",
  y_var = "y_median",
  y_lower = "y_quant_25",
  y_upper = "y_quant_75",
  y_min = "y_quant_5",
  y_max = "y_quant_95",
  arrange_by = "total",
  arrange_desc = FALSE,
  add_total = TRUE,
  horizontal = TRUE
)

box_plot(
  df = df,
  x_var = "unit",
  y_var = "y_median",
  y_lower = "y_quant_25",
  y_upper = "y_quant_75",
  y_min = "y_quant_5",
  y_max = "y_quant_95",
  arrange_by = "total",
  arrange_desc = TRUE,
  add_total = TRUE,
  horizontal = TRUE
)
}
