% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors_rc}
\alias{colors_rc}
\title{Wrapper for RColorBrewer to get colors}
\usage{
colors_rc(n = 3, type = "default")
}
\arguments{
\item{n}{number of colors}

\item{type}{one of
\code{"default"}, for standard colors.
\code{"RYG"}, for red-yellow-green scale.
\code{"Paired"} for pairwise colors.}
}
\value{
vector with color codes
}
\description{
Wrapper for RColorBrewer to get colors
}
\examples{
colors_rc(2)
}
