% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors_rc_2}
\alias{colors_rc_2}
\title{Registercentrum standard colors}
\usage{
colors_rc_2(n, type = "qualitative")
}
\arguments{
\item{n}{the number of colors desired}

\item{type}{the type of palette, one of "qualitative", "sequential", and
"diverging"}
}
\description{
Get a number of colors according to Registercentrum standards. Three palettes
are available based on the type of palette desired.
}
\details{
A qualitative palette should be used when we have groups that are all equal
in value, for example some different type of treatment. A sequential palette
should be used when data has an inherent order, for example age groups, pain
scale, etc. Finally, a diverging palette should be used when data is centered
around a neutral value, for example data measuring improvement/deteriation
after a certain treatment.

The qualitative palette is Okabe-Ito, the sequential is viridis, and the
divering is Purple-Green.
}
