% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors_rc_3}
\alias{colors_rc_3}
\title{Registercentrum standard colors}
\usage{
colors_rc_3(n, type = "qualitative")
}
\arguments{
\item{n}{the number of colors desired}

\item{type}{the type of palette, one of "qualitative", "sequential_1", and
"sequential_2"}
}
\description{
Color palettes used as a standard in Registercentrum.
The colors are based on input from UX-designers and have been chosen to
accommodate levels of visual impairment.
}
\details{
There are 3 color palettes, categorical (used for grouped data in
which the categories are not supposed to be graded) and two sequential
palletes.
}
