% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swe_data.R
\docType{data}
\name{counties}
\alias{counties}
\title{Map coordinates for Sweden counties}
\format{
A sf dataset with one row per county
\describe{
\item{id}{ID of county}
\item{geometry}{Coordinate polygon used to draw county borders}
}
}
\source{
\url{http://api.thenmap.net/v2/se-4/geo/2020-06-03}
}
\usage{
counties
}
\description{
Map coordinates for Sweden counties
}
\examples{
library(sf)
library(ggplot2)
ggplot() +
 geom_sf(data = counties, aes(fill = as.numeric(id))) +
 theme_void()

ggplot() +
  geom_sf(data = municipalities, aes(fill = as.numeric(id))) +
  geom_sf(data = counties, fill = NA, color = "red") +
 theme_void()

}
\keyword{datasets}
