% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{density_plot}
\alias{density_plot}
\title{Produce a density plot}
\usage{
density_plot(
  df,
  var,
  color_var = NULL,
  counted = FALSE,
  count_var = NULL,
  alpha = 0.2,
  adjust = 1.25,
  y_breaks = ggplot2::waiver(),
  y_labels = scales::label_percent(),
  y_lim = NULL,
  x_breaks = ggplot2::waiver(),
  x_labels = ggplot2::waiver(),
  x_lim = NULL,
  x_lab = NULL,
  y_lab = NULL,
  title = ggplot2::waiver(),
  palette_type = "qualitative",
  color_title = NULL,
  plotly = FALSE,
  facet = FALSE,
  facet_by = NULL,
  colors = NULL
)
}
\arguments{
\item{df}{data.frame to plot}

\item{var}{the name of the variable to use when creating density}

\item{color_var}{the name of the color variable}

\item{counted}{whether or not the data.frame is already counted, typically
the result of \code{\link[dplyr:count]{dplyr::count()}}}

\item{count_var}{the name of the count variablet, typically "n"}

\item{alpha}{alpha passed to \code{\link[ggplot2:geom_density]{ggplot2::geom_density()}}}

\item{adjust}{the bandwidth used when creating the density, passed to
\code{\link[ggplot2:geom_density]{ggplot2::geom_density()}}}

\item{y_breaks}{vector of breaks for y-axis}

\item{y_labels}{vector of labels for y-labels}

\item{y_lim}{limits for the y-axis}

\item{x_breaks}{vector of breaks for x-axis}

\item{x_labels}{vector of labels for x-axis}

\item{x_lim}{limits for the x-axis}

\item{x_lab}{the label for the x-axis}

\item{y_lab}{the label for the y-axis}

\item{title}{the title of the plot}

\item{palette_type}{the type of palette, one of "qualitative", "sequential",
and "diverging"}

\item{color_title}{the title of the color legend}

\item{plotly}{whether or not to convert the finished plot to a plotly
graph}

\item{facet}{whether or not to facet the plots using \link[ggplot2:facet_wrap]{ggplot2::facet_wrap}}

\item{facet_by}{the name of the variable to facet by}

\item{colors}{manually specified colors to use for fill, only used when
\code{palette_type} is \code{qualitative}, must be subset of
\code{colors_rc_2(9, "qualitative")}}
}
\description{
Produces a density plot according to Registercentrum graphical standards
}
\details{
If the data.frame is already counted it will be uncounted using
\link[tidyr:uncount]{tidyr::uncount}.

The arguments \verb{*_lim}, \verb{*_breaks}, and \verb{*_labels} are passed to
\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous} and \link[ggplot2:scale_discrete]{ggplot2::scale_x_discrete}.

The arguments \code{x_lab}, \code{y_lab}, and \code{title} are passed to \code{\link[ggplot2:labs]{ggplot2::labs()}}.
}
\examples{
df <- data.frame(
  x = c(rnorm(5000, mean = 4), rnorm(5000, mean = 5)),
  sex = rep(1:2, each = 5000)
)

density_plot(
  df = df,
  var = "x"
)

density_plot(
  df = df,
  var = "x",
  color_var = "sex",
  color_title = "Kön",
  y_lab = "Density",
  x_lab = "x title"
)

density_plot(
  df = df,
  var = "x",
  color_var = "sex",
  color_title = "Kön",
  y_lab = "Density",
  x_lab = "x title",
  plotly = TRUE
)

df_2 <- data.frame(
  age = 18:30,
  n = 18:30 * 1000
)

density_plot(
  df = df_2,
  var = "age",
  counted = TRUE,
  count_var = "n",
  adjust = 2,
  x_lim = c(10, 40)
)

density_plot(
  df = df_2,
  var = "age",
  counted = TRUE,
  count_var = "n",
  adjust = 2,
  x_lim = c(10, 40),
  x_lab = "Ålder",
  title = "Hej",
  plotly = TRUE
)

df_3 <- data.frame(
  x = c(rnorm(5000, mean = 4), rnorm(5000, mean = 5)),
  sex = rep(1:2, each = 5000),
  unit = letters[1:2]
)

density_plot(
  df = df_3,
  var = "x",
  color_var = "sex",
  color_title = "Kön",
  facet = TRUE,
  facet_by = "unit"
)
}
