% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr_data.R
\docType{data}
\name{eq_vas}
\alias{eq_vas}
\alias{eq_vas_riket}
\alias{eq_vas_obs}
\alias{eq_vas_exp}
\title{Swedish average EQ VAS dataset}
\format{
A data frame with 1 row and 5 variables:
\describe{
\item{hospital}{hospital name, character}
\item{r2008_2009}{Swedish average EQ VAS for 2008/2009, numeric}
\item{r2010_2011}{Swedish average EQ VAS for 2010/2011, numeric}
\item{r2012_2013}{Swedish average EQ VAS for 2012/2013, numeric}
\item{r2014_2015}{Swedish average EQ VAS for 2014/2015, numeric}
}

A data frame with 68 rows and 5 variables: \describe{
\item{hospital}{hospital name, character} \item{eqvas2008_2009}{observed EQ
VAS mean for 2008/2009, numeric} \item{eqvas2010_2011}{observed EQ VAS mean
for 2010/2011, numeric} \item{eqvas2012_2013}{observed EQ VAS mean for
2012/2013, numeric} \item{eqvas2014_2015}{observed EQ VAS mean for
2014/2015, numeric} }

A data frame with 68 rows and 5 variables: \describe{
\item{hospital}{hospital name, character} \item{res2008_2009}{expected EQ
VAS mean for 2008/2009, numeric} \item{res2010_2011}{expected EQ VAS mean
for 2010/2011, numeric} \item{res2012_2013}{expected EQ VAS mean for
2012/2013, numeric} \item{res2014_2015}{expected EQ VAS mean for 2014/2015,
numeric} }
}
\usage{
eq_vas_riket

eq_vas_obs

eq_vas_exp
}
\description{
A dataset containing Swedish average EQ VAS for four year periods:
2008/2009, 2010/2011, 2012/2013, and 2014/2015. Data set used for
prom_trends() examples.

A dataset containing observed EQ VAS for four year periods: 2008/2009,
2010/2011, 2012/2013, and 2014/2015. Data set used for prom_trends()
examples.

A dataset containing expected EQ VAS calculated with a linear model for four
year periods: 2008/2009, 2010/2011, 2012/2013, and 2014/2015. Data set used
for prom_trends() examples.
}
\keyword{EQ}
\keyword{SLR}
\keyword{VAS}
