% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{export_highcharts}
\alias{export_highcharts}
\title{Export a highcharts config}
\usage{
export_highcharts(cfg, write_clip = TRUE)
}
\arguments{
\item{cfg}{highcharts config}

\item{write_clip}{should result be saved to clipboard}
}
\value{
the resulting JSON character, invisibly if write_clip is TRUE
}
\description{
Converts a highcharts config to a JSON ready to paste in the highcharts
sandbox
}
