% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{get_size_params}
\alias{get_size_params}
\title{Create parameters for sizing and padding for bar plot.}
\usage{
get_size_params(out, position)
}
\arguments{
\item{out}{config}

\item{position}{if bar should be stacked or dodge}
}
\description{
Function for getting the values of the bar height, bar width,
group padding etc.
}
