% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{get_text_size}
\alias{get_text_size}
\title{Get text size param for plots}
\usage{
get_text_size(out, text_size = 14)
}
\arguments{
\item{out}{config}

\item{text_size}{size of text, will be interperted as pixel}
}
\description{
Get text size param for plots
}
