% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_plot.R
\name{km_plot_2}
\alias{km_plot_2}
\title{Title}
\usage{
km_plot_2(
  survfit_obj,
  n_risk_break = 50,
  x_lab = NULL,
  y_lab = "Survival probablity",
  y_breaks = ggplot2::waiver(),
  y_labels = scales::label_percent(),
  y_lim = c(0, 1),
  title = ggplot2::waiver(),
  x_breaks = ggplot2::waiver(),
  x_labels = ggplot2::waiver(),
  x_lim = NULL,
  add_ci = TRUE,
  ci_alpha = 0.2,
  risk_table = TRUE,
  plot_type = NULL,
  legend_title = "Strata",
  legend_position = "top",
  legend_labs = NULL,
  facet_by = NULL,
  plotly = FALSE
)
}
\arguments{
\item{survfit_obj}{an object of class survfit, typically produced by
\code{\link[survival:survfit]{survival::survfit()}}}

\item{n_risk_break}{when the number of subjects at risk is lower than this
argument the survival probability is censored from there on}

\item{x_lab}{the label for the x-axis}

\item{y_lab}{the label for the y-axis}

\item{y_breaks}{vector of breaks for y-axis}

\item{y_labels}{vector of labels for y-labels}

\item{y_lim}{limits for y-axis}

\item{title}{the title of the plot}

\item{x_breaks}{vector of breaks for x-axis}

\item{x_labels}{vector of labels for x-axis}

\item{x_lim}{limits x-axis}

\item{add_ci}{whether or not to add a confidence interval}

\item{ci_alpha}{the alpha of the confidence interval ribbon}

\item{risk_table}{whether or not to add a table describing the number at risk}

\item{plot_type}{one of NULL (for standard survival), "cumhaz" (for
cumulative hazard), and "event" (for cumulative events)}

\item{legend_title}{the title of the legend}

\item{legend_position}{one of "top", "bottom", "left", "right", and "none}

\item{legend_labs}{the labels for the legend}

\item{facet_by}{the name of the variable to facet by}

\item{plotly}{whether or not to convert the finished plot to a plotly
graph}
}
\description{
Title
}
\examples{
fit <- survival::survfit(
  survival::Surv(time / 365.24, status) ~ sex,
  survival::colon
)

km_plot_2(fit)

km_plot_2(
  survfit_obj = fit,
  add_ci = FALSE,
  x_lab = "time",
  y_lab = "surv",
  legend_title = "legend title",
  legend_labs = c("man", "woman"),
  legend_position = "right"
)

km_plot_2(
  fit,
  plot_type = "cumhaz"
)

fit <- survival::survfit(
  survival::Surv(time / 365.24, status) ~ sex + rx,
  survival::colon
)

km_plot_2(
  fit,
  facet_by = "rx"
)

km_plot_2(
  fit
)

km_plot_2(
  fit,
  facet_by = "rx",
  risk_table = TRUE,
  plotly = TRUE
)

fit_3 <- survival::survfit(
  survival::Surv(time / 365.24, status) ~ 1,
  survival::colon
)

km_plot_2(fit_3)
}
