% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot.R
\name{line_plot}
\alias{line_plot}
\title{Line plot function}
\usage{
line_plot(
  df,
  x_var,
  color_var = NULL,
  y_var = NULL,
  group_by_x_var = TRUE,
  y_percent = TRUE,
  percent_accuracy = 1,
  y_lim = NULL,
  y_breaks = 2000,
  y_breaks_end = 1e+05,
  line_size = 1,
  title = NULL,
  subtitle = NULL,
  y_lab = NULL,
  x_lab = NULL,
  fill_colors = NULL,
  legend_labels = ggplot2::waiver(),
  label_breaks = ggplot2::waiver(),
  legend_row = NULL,
  legend_col = NULL,
  expand = TRUE,
  ...
)
}
\arguments{
\item{df}{Data frame.}

\item{x_var}{Variable for x-axis, use string name.
Recommended that x_var is
in character in df (not necessary).}

\item{color_var}{Variable for the different colors in lines, use
string name. Use \code{NULL} if only one color for
lines.}

\item{y_var}{Variable for y axis, if \code{NULL}, count is used.}

\item{group_by_x_var}{Boolean indicating if percentages should be for
\code{x_var} or \code{color_var}.}

\item{y_percent}{If \code{TRUE}, y-axis is in percent form. Otherwise in
count form.}

\item{percent_accuracy}{Set accuracy for \code{\link[scales:percent_format]{scales::percent_format()}}.}

\item{y_lim}{Limit on y-axis.}

\item{y_breaks}{Length between each break on y-axis.}

\item{y_breaks_end}{Break end, default for 100,000. Works for all count
values below that.}

\item{line_size}{Size of the lines.}

\item{title}{Plot title, \code{NULL} if no title.}

\item{subtitle}{Small text under title, \code{NULL} if no subtitle.}

\item{y_lab}{Y-axis label, use \code{NULL} for no label.}

\item{x_lab}{X-axis label, use \code{NULL} for no label.}

\item{fill_colors}{Colors of the different categories in color_var.}

\item{legend_labels}{Label for each legend key.}

\item{label_breaks}{Order of the legend keys.}

\item{legend_row}{How many rows for the legends.}

\item{legend_col}{How many columns for the legends.}

\item{expand}{If \code{TRUE}, the margins around the data are kept.}

\item{...}{Arguments passed to \code{\link[=theme_rc]{theme_rc()}}}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{            Ggplot object containing line-plot.
}\if{html}{\out{</div>}}
}
\description{
Standard line plot using ggplot2. Y-variable not necessary.
}
