% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot.R
\name{line_plot_2}
\alias{line_plot_2}
\title{Produce a line plot}
\usage{
line_plot_2(
  df,
  x_var,
  y_var,
  linetype_var = NULL,
  color_var = NULL,
  y_breaks = ggplot2::waiver(),
  y_labels = ggplot2::waiver(),
  y_lim = NULL,
  x_breaks = ggplot2::waiver(),
  x_labels = ggplot2::waiver(),
  x_lim = NULL,
  point = TRUE,
  x_lab = NULL,
  y_lab = NULL,
  title = ggplot2::waiver(),
  palette_type = "qualitative",
  linetype_title = NULL,
  color_title = NULL,
  plotly = FALSE,
  facet = FALSE,
  facet_by = NULL,
  color_nrow = 1,
  linetype_nrow = 1,
  point_shape_var = NULL,
  colors = NULL,
  remove_grid = TRUE,
  text_size = 7
)
}
\arguments{
\item{df}{data.frame to plot}

\item{x_var}{the name of the x variable}

\item{y_var}{the name of the y variable}

\item{linetype_var}{the name of the linetype variable}

\item{color_var}{the name of the color variable}

\item{y_breaks}{vector of breaks for y-axis}

\item{y_labels}{vector of labels for y-labels}

\item{y_lim}{limits for the y-axis}

\item{x_breaks}{vector of breaks for x-axis}

\item{x_labels}{vector of labels for x-axis}

\item{x_lim}{limits for the x-axis}

\item{point}{whether or not to add points}

\item{x_lab}{the label for the x-axis}

\item{y_lab}{the label for the y-axis}

\item{title}{the title of the plot}

\item{palette_type}{the type of palette, one of "qualitative", "sequential",
and "diverging"}

\item{linetype_title}{the title of the linetype legend}

\item{color_title}{the title of the color legend}

\item{plotly}{whether or not to convert the finished plot to a plotly
graph}

\item{facet}{whether or not to facet the plots using \link[ggplot2:facet_wrap]{ggplot2::facet_wrap}}

\item{facet_by}{the name of the variable to facet by}

\item{color_nrow}{the number of rows for the color legend}

\item{linetype_nrow}{the number of rows for the linetype legend}

\item{point_shape_var}{the name of the variable used to adjust the shapes of
the points if \code{point} is \code{TRUE}}

\item{colors}{manually specified colors to use for fill, only used when
\code{palette_type} is \code{qualitative}, must be subset of
\code{colors_rc_2(9, "qualitative")}}

\item{remove_grid}{if gridlines should be removed}

\item{text_size}{passed to \code{\link[=theme_rc]{theme_rc()}}}
}
\description{
Produces a line plot according to Registercentrum graphical standards.
}
\details{
The arguments \verb{*_lim}, \verb{*_breaks}, and \verb{*_labels} are passed to
\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous} and \link[ggplot2:scale_discrete]{ggplot2::scale_x_discrete}.

The arguments \code{x_lab}, \code{y_lab}, and \code{title} are passed to \code{\link[ggplot2:labs]{ggplot2::labs()}}.
}
\examples{
df <- data.frame(
  year = 2010:2020,
  prop = pmin(rnorm(n = 11, mean = 0.8, sd = 0.1), 1),
  n = rpois(11, 25)
)

line_plot_2(
  df = df,
  x_var = "year",
  y_var = "prop",
  y_labels = scales::label_percent(),
  x_breaks = 2010:2020,
  y_lim = c(0, 1)
)

line_plot_2(
  df = df,
  x_var = "year",
  y_var = "prop",
  y_labels = scales::label_percent(),
  x_breaks = 2010:2020,
  y_lim = c(0, 1),
  plotly = TRUE
)


df_2 <- data.frame(
  year = rep(2010:2020, each = 4),
  some_var = rep(rep(c("prop", "rapporteringsgrad"), each = 2), 11),
  county = c("a", "Riket"),
  prop = pmin(rnorm(n = 22, mean = 0.8, sd = 0.1), 1),
  n = rpois(11, 25)
)

line_plot_2(
  df = df_2,
  x_var = "year",
  y_var = "prop",
  linetype_var = "some_var",
  color_var = "county",
  y_labels = scales::label_percent(),
  x_breaks = 2010:2020,
  color_title = "Region",
  x_lab = "År",
  y_lab = "Andel (\%)",
  linetype_title = "Linetype"
)

line_plot_2(
  df = df_2,
  x_var = "year",
  y_var = "prop",
  linetype_var = "some_var",
  color_var = "county",
  y_labels = scales::label_percent(),
  x_breaks = 2010:2020,
  color_title = "Region",
  plotly = TRUE,
  x_lab = "År",
  y_lab = "Andel (\%)",
  linetype_title = "Linetype"
)

line_plot_2(
  df = df_2,
  x_var = "year",
  y_var = "prop",
  linetype_var = "some_var",
  facet = TRUE,
  facet_by = "county"
)
}
