% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swe_data.R
\docType{data}
\name{municipalities}
\alias{municipalities}
\title{Map coordinates for Sweden municipalites}
\format{
A sf dataset with one row per municipality
\describe{
\item{id}{ID of municipality with region code (as in data from SCB)}
\item{geometry}{Coordinate polygon used to draw municipality borders}
\item{Name}{Name of municipality}
\item{RegionID}{County code}
\item{MunicipID}{Municipality code}
}
}
\source{
\url{http://api.thenmap.net/v2/se-7/geo/2020-06-03}
\url{"https://www.scb.se/contentassets/7a89e48960f741e08918e489ea36354a/kommunlankod_20211229.xls"}
}
\usage{
municipalities
}
\description{
Map coordinates for Sweden municipalites
}
\examples{
# Use internal coordinate datasets
region_coords <- counties
municip_coords <- municipalities

# Map with labels

map_plot(
  df = region_coords,
  fill_var = "id",
  label_var = "Name"
)

# Add some "interresting" data to coordinate dataset and plot
region_coords[["var"]] <- sample(c("Bäst", "Bra", "Dålig", "Sämst"),
                                 size = 21, replace = TRUE)

map_plot(
  df = region_coords,
  fill_var = "var"
)

# Plot both municipalities and region outlines
ggplot2::ggplot() +
  ggplot2::geom_sf(data = municipalities, ggplot2::aes(fill = as.numeric(id))) +
  ggplot2::geom_sf(data = counties, fill = NA, color = "red") +
  ggplot2::theme_void()

# Plot municipalities of Västra Götalands Län, with and without labels
vgregion <- dplyr::filter(municipalities, RegionID == "14")

vgregion[["var"]] <- sample(c("Bäst", "Bra", "Dålig", "Sämst"),
                            size = nrow(vgregion), replace = TRUE)

map_plot(
  df = vgregion,
  fill_var = "var"
)
map_plot(
  df = vgregion,
  fill_var = "var",
  label_var = "Name"
)


}
\keyword{datasets}
