% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{ndr_colors}
\alias{ndr_colors}
\title{NDR color palette}
\usage{
ndr_colors(n = 8, type = NULL)
}
\arguments{
\item{n}{number of colors}

\item{type}{one of
\code{NULL} for default
\code{"deep"}, for dark colors.
\code{"gyr"}, for scale green to yellow to red.
\code{"Paired"} for pairwise colors.}
}
\value{
vector with color codes
}
\description{
NDR color palette
}
\examples{
ndr_colors(2)
}
