% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{order_x_var}
\alias{order_x_var}
\title{Order the x variable}
\usage{
order_x_var(df, x_var = NULL, order = NULL)
}
\arguments{
\item{df}{data.frame}

\item{x_var}{the x variable}

\item{order}{vector/list of values giving the desired level order}
}
\value{
data.frame
}
\description{
Converts \code{x_var} to a factor with levels in the supplied order
}
