% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr_data.R
\docType{data}
\name{pain_vas}
\alias{pain_vas}
\alias{pain_riket}
\alias{pain_obs}
\alias{pain_exp}
\title{Swedish average pain VAS datasets}
\format{
A data frame with 1 row and 5 variables: \describe{
\item{hospital}{hospital name, character} \item{r2008_2009}{Swedish average
pain VAS for 2008/2009, numeric} \item{r2010_2011}{Swedish average pain VAS
for 2010/2011, numeric} \item{r2012_2013}{Swedish average pain VAS for
2012/2013, numeric} \item{r2014_2015}{Swedish average pain VAS for
2014/2015, numeric} }

A data frame with 68 rows and 5 variables: \describe{
\item{hospital}{hospital name, character} \item{pain2008_2009}{observed
pain VAS mean for 2008/2009, numeric} \item{pain2010_2011}{observed pain
VAS mean for 2010/2011, numeric} \item{pain2012_2013}{observed pain VAS
mean for 2012/2013, numeric} \item{pain2014_2015}{observed pain VAS mean
for 2014/2015, numeric} }

A data frame with 68 rows and 5 variables: \describe{
\item{hospital}{hospital name, character} \item{res2008_2009}{expected pain
VAS mean for 2008/2009, numeric} \item{res2010_2011}{expected pain VAS mean
for 2010/2011, numeric} \item{res2012_2013}{expected pain VAS mean for
2012/2013, numeric} \item{res2014_2015}{expected pain VAS mean for
2014/2015, numeric} }
}
\usage{
pain_riket

pain_obs

pain_exp
}
\description{
datasets containing Swedish average pain VAS for four year periods:
2008/2009, 2010/2011, 2012/2013, and 2014/2015. Data set used for
prom_trends() examples.

A dataset containing observed pain VAS for four year periods: 2008/2009,
2010/2011, 2012/2013, and 2014/2015. Data set used for prom_trends()
examples.

A dataset containing expected pain VAS calculated with a linear model for
four year periods: 2008/2009, 2010/2011, 2012/2013, and 2014/2015. Data set
used for prom_trends() examples.
}
\keyword{SLR}
\keyword{VAS}
\keyword{pain}
