% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{plot_highcharts}
\alias{plot_highcharts}
\alias{add_y_axis}
\alias{add_tooltip}
\alias{make_series}
\title{Create a config for a highchart plot}
\usage{
plot_highcharts(
  df,
  x_var,
  vars,
  title,
  group_vars,
  y_lim,
  y_breaks,
  x_breaks = NULL,
  proportion = FALSE,
  scale_percentage = FALSE,
  type,
  other_vars,
  horizontal = FALSE,
  x_lab = NULL,
  y_lab = NULL,
  arrange_by = NULL,
  arrange_desc = TRUE,
  arrange_by_group_var = NULL,
  group_var_order = NULL,
  plot_height = 0.8,
  group_color = lifecycle::deprecated(),
  legend_title = NULL,
  reversed_stacks = NULL,
  text_size = NULL,
  remove_value = NULL,
  horizontal_line = NULL,
  n_decimals = 0
)

add_y_axis(
  out,
  y_lim = NULL,
  y_breaks = NULL,
  proportion = FALSE,
  y_lab = NULL,
  reversed_stacks = NULL,
  horizontal_line = NULL,
  scale_percentage = TRUE
)

add_tooltip(
  out,
  df,
  group_vars = NULL,
  other_vars = NULL,
  proportion = FALSE,
  type,
  x_var
)

make_series(
  df,
  vars,
  group_vars = NULL,
  palette_type = "qualitative",
  colors = lifecycle::deprecated(),
  other_vars = NULL,
  proportion = FALSE,
  scale_percentage = TRUE,
  group_var_order = NULL,
  x_var,
  n_decimals = 0
)
}
\arguments{
\item{df}{the data.frame to plot}

\item{x_var}{x variable}

\item{vars}{a named list used to rename variables to proper highcharts keys,
typically \code{list(y = y_var)}}

\item{title}{the title of plot}

\item{group_vars}{the variables indicating a series to plot}

\item{y_lim}{limits for the y-axis}

\item{y_breaks}{vector of breaks for y-axis}

\item{x_breaks}{vector of breaks for x-axis}

\item{proportion}{whether the y variable is a proportion}

\item{scale_percentage}{whether or not to scale the y-variable by 100}

\item{type}{the type of graph}

\item{other_vars}{other variables to include in the tooltip, should be a
named list where the name will be the key in the tooltip}

\item{horizontal}{if plot should be horizontal}

\item{x_lab}{the title of the x axis}

\item{y_lab}{the title of the y axis}

\item{arrange_by}{column to sort by}

\item{arrange_desc}{if sort is descending}

\item{arrange_by_group_var}{value in group_var to sort by}

\item{group_var_order}{order of group var, can
alternatively be \code{auto_character} or \code{auto_numeric} to automatically sort
the levels}

\item{plot_height}{height of plot}

\item{group_color}{deprecated}

\item{legend_title}{title of the legend}

\item{reversed_stacks}{should stacks be reversed?}

\item{text_size}{size of text, will be interperted as pixels}

\item{remove_value}{value to remove, useful for removing obfuscated
observations}

\item{horizontal_line}{horizontal reference line}

\item{n_decimals}{number of decimals to round numbers to}

\item{out}{intermediate config}

\item{palette_type}{passed to \code{\link[=colors_rc_3]{colors_rc_3()}}}

\item{colors}{deprecated}
}
\value{
highcharts config
}
\description{
Create a config for a highchart plot
}
\section{Functions}{
\itemize{
\item \code{add_y_axis()}: add the y axis to the config

\item \code{add_tooltip()}: add the tooltip to the config

\item \code{make_series()}: converts a data.frame to a highcharts series
representation

}}
