% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plots.R
\name{plotly_to_html}
\alias{plotly_to_html}
\title{Converts a plotly object to an HTML character}
\usage{
plotly_to_html(plt)
}
\arguments{
\item{plt}{a plotly object}
}
\value{
an HTML character
}
\description{
Converts a plotly object to an HTML character
}
