% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar_plot.R
\name{radar_plot}
\alias{radar_plot}
\title{Produce a radar plot}
\usage{
radar_plot(
  df,
  seg = NULL,
  palette_type = "qualitative",
  y_lim = NULL,
  legend_col = NULL
)
}
\arguments{
\item{df}{A data frame containing the values for the radar plot.}

\item{seg}{Number of segments in the radar chart. If \code{NULL}, it is
automatically set based on \code{pretty(y_min:y_max)}.}

\item{palette_type}{selects the type of colors used for the groups in the
chart can be one of "qualitative", "sequential" or "diverging"}

\item{y_lim}{Minimum and maximum value for the radar chart axes. If \code{NULL},
it is automatically set to 0 - the maximum value in the df + 1}

\item{legend_col}{The name of the column containing group names for the
legend.}
}
\value{
A radar chart plotted using \code{fmsb::radarchart()}.
}
\description{
Produces a radar plot using \code{fmsb::radarchart()}, with automatic handling
of axis limits and colors.
}
\examples{
df <- data.frame(A = c(3, 4, 5), B = c(2, 3, 4), C = c(4, 5, 3))
radar_plot(df)
}
