% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcplot-package.R
\docType{package}
\name{rcplot-package}
\alias{rcplot}
\alias{rcplot-package}
\title{rcplot: Plot functions for Registercentrum}
\description{
Implementing the graphical design used by Registercentrum. Used for annual reports and more.
}
\author{
\strong{Maintainer}: Jonatan Nåtman \email{jonatan.natman@vgregion.se}

Authors:
\itemize{
  \item Rikard Isaksson \email{rikard.isaksson@vgregion.se}
  \item Erik Bülow \email{erik.bulow@vgregion.se}
}

}
\keyword{internal}
