% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rcsave}
\alias{rcsave}
\title{ggsave function with default settings as specified with options()}
\usage{
rcsave(
  file = NULL,
  plot = ggplot2::last_plot(),
  width = ifelse(!is.null(options("figure_width")), options("figure_width"), 3.14961),
  height = ifelse(!is.null(options("figure_height")), options("figure_height"),
    3.54331),
  units = "in",
  dpi = ifelse(!is.null(options("figure_dpi")), options("figure_dpi"), 600),
  device = ifelse(!is.null(options("figure_device")), options("figure_device"), "tiff"),
  family = ifelse(!is.null(options("figure_family")), options("figure_family"),
    "Calibri Light")
)
}
\arguments{
\item{file}{File name to create on disk.}

\item{plot}{Plot to save}

\item{width}{Width in units ("in", "cm", or "mm").
default: options("figure_width")}

\item{height}{Height in units ("in", "cm", or "mm").}

\item{units}{Units ("in", "cm", or "mm")}

\item{dpi}{Plot resolution.}

\item{device}{Device to use. Can either be a device function (e.g. png()), or one of "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{family}{Font family}
}
\description{
ggsave function with default settings as specified with options()
}
