% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_rc}
\alias{theme_rc}
\title{The default theme for Registercentrum}
\usage{
theme_rc(
  axis_text_angle = NULL,
  legend_position = "bottom",
  legend_justification = legend_position,
  text_size = 7,
  subtitle_size = 8,
  title_size = 9,
  title_hjust = 0.5,
  subtitle = FALSE,
  title_margin = if (subtitle) 1 else title_size/2,
  remove_grid = TRUE,
  remove_legend = FALSE
)
}
\arguments{
\item{axis_text_angle}{passed to
\code{axis.text.x = element_text(angle = axis_text_angle)}}

\item{legend_position}{passed to legend.posision}

\item{legend_justification}{passed to legend.justification}

\item{text_size, title_size, subtitle_size}{Text size for most text,
title and subtitle}

\item{title_hjust, title_margin}{passed to
\verb{plot.title = element_text(hjust = title_hjust, margin = margin(b = title_margin)}}

\item{subtitle}{is a subtitle used in the figure (boolean)?}

\item{remove_grid}{If grid should be added}

\item{remove_legend}{If legends should be removed}

\item{plot_type}{One of "bar" or "line" controling the major grid lines}
}
\value{
Modified version of \code{\link[=theme_classic]{theme_classic()}}
}
\description{
Implementation of graphical profile used in package by
default.
}
