% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr_data.R
\docType{data}
\name{satisfaction_vas}
\alias{satisfaction_vas}
\alias{satis_riket}
\alias{satis_obs}
\alias{satis_exp}
\title{Swedish average satisfaction VAS dataset}
\format{
A data frame with 1 row and 5 variables: \describe{
\item{hospital}{hospital name, character} \item{r2008_2009}{Swedish average
satisfaction VAS for 2008/2009, numeric} \item{r2010_2011}{Swedish average
satisfaction VAS for 2010/2011, numeric} \item{r2012_2013}{Swedish average
satisfaction VAS for 2012/2013, numeric} \item{r2014_2015}{Swedish average
satisfaction VAS for 2014/2015, numeric} }

A data frame with 68 rows and 5 variables: \describe{
\item{hospital}{hospital name, character} \item{sati2008_2009}{observed
satisfaction VAS mean for 2008/2009, numeric} \item{sati2010_2011}{observed
satisfaction VAS mean for 2010/2011, numeric} \item{sati2012_2013}{observed
satisfaction VAS mean for 2012/2013, numeric} \item{sati2014_2015}{observed
satisfaction VAS mean for 2014/2015, numeric} }

A data frame with 68 rows and 5 variables: \describe{
\item{hospital}{hospital name, character} \item{res2008_2009}{expected
satisfaction VAS mean for 2008/2009, numeric} \item{res2010_2011}{expected
satisfaction VAS mean for 2010/2011, numeric} \item{res2012_2013}{expected
satisfaction VAS mean for 2012/2013, numeric} \item{res2014_2015}{expected
satisfaction VAS mean for 2014/2015, numeric} }
}
\usage{
satis_riket

satis_obs

satis_exp
}
\description{
A dataset containing Swedish average satisfaction VAS for four year periods:
2008/2009, 2010/2011, 2012/2013, and 2014/2015. Data set used for
prom_trends() examples.

A dataset containing observed satisfaction VAS for four year periods:
2008/2009, 2010/2011, 2012/2013, and 2014/2015. Data set used for
prom_trends() examples.

A dataset containing expected satisfaction VAS calculated with a linear model
for four year periods: 2008/2009, 2010/2011, 2012/2013, and 2014/2015. Data
set used for prom_trends() examples.
}
\keyword{SLR}
\keyword{satisfaction-VAS}
