% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Produce a scatter plot}
\usage{
scatter_plot(
  df,
  x_var,
  y_var,
  color_var = NULL,
  shape_var = NULL,
  x_breaks = ggplot2::waiver(),
  x_labels = ggplot2::waiver(),
  x_lim = NULL,
  y_breaks = ggplot2::waiver(),
  y_labels = ggplot2::waiver(),
  y_lim = NULL,
  x_lab = ggplot2::waiver(),
  y_lab = ggplot2::waiver(),
  title = ggplot2::waiver(),
  palette_type = "qualitative",
  color_title = NULL,
  shape_title = NULL,
  plotly = FALSE,
  facet = FALSE,
  facet_by = NULL,
  colors = NULL
)
}
\arguments{
\item{df}{data.frame to plot}

\item{x_var}{the name of the x variable}

\item{y_var}{the name of the y variable}

\item{color_var}{the name of the color variable}

\item{shape_var}{the name of the shape variable}

\item{x_breaks}{vector of breaks for x-axis}

\item{x_labels}{vector of labels for x-axis}

\item{x_lim}{limits for the x-axis}

\item{y_breaks}{vector of breaks for y-axis}

\item{y_labels}{vector of labels for y-labels}

\item{y_lim}{limits for the y-axis}

\item{x_lab}{the label for the x-axis}

\item{y_lab}{the label for the y-axis}

\item{title}{the title of the plot}

\item{palette_type}{the type of palette, one of "qualitative", "sequential",
and "diverging"}

\item{color_title}{the title of the color legend}

\item{shape_title}{the title of the shape legend}

\item{plotly}{whether or not to convert the finished plot to a plotly
graph}

\item{facet}{whether or not to facet the plots using \link[ggplot2:facet_wrap]{ggplot2::facet_wrap}}

\item{facet_by}{the name of the variable to facet by}

\item{colors}{manually specified colors to use for fill, only used when
\code{palette_type} is \code{qualitative}, must be subset of
\code{colors_rc_2(9, "qualitative")}}
}
\description{
Produce a scatter plot according to Registercentrum graphical standards
}
\details{
The arguments \verb{*_lim}, \verb{*_breaks}, and \verb{*_labels} are passed to
\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous} and \link[ggplot2:scale_discrete]{ggplot2::scale_x_discrete}.

The arguments \code{x_lab}, \code{y_lab}, and \code{title} are passed to \code{\link[ggplot2:labs]{ggplot2::labs()}}.
}
\examples{
df <- data.frame(
  x = c(rnorm(500), rnorm(500, 2)),
  y = c(rnorm(500), rnorm(500, 2)),
  sex = rep(1:2, each = 500),
  age_cat = rep(c("<18", ">=18"), each = 2, size = 500)
)

scatter_plot(
  df = df,
  x_var = "x",
  y_var = "y"
)

scatter_plot(
  df = df,
  x_var = "x",
  y_var = "y",
  color_var = "sex",
  color_title = "Kön"
)

scatter_plot(
  df = df,
  x_var = "x",
  y_var = "y",
  color_var = "sex",
  shape_var = "age_cat"
)

scatter_plot(
  df = df,
  x_var = "x",
  y_var = "y",
  color_var = "sex",
  shape_var = "age_cat",
  x_lab = "En x-variabel",
  y_lab = "En y-variabel",
  plotly = TRUE,
  color_title = "Kön",
  shape_title = "Ålderskategori"
)

scatter_plot(
  df = df,
  x_var = "x",
  y_var = "y",
  plotly = TRUE
)

scatter_plot(
  df = df,
  x_var = "x",
  y_var = "y",
  facet = TRUE,
  facet_by = "sex"
)
}
