% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{set_text_size}
\alias{set_text_size}
\title{Get text size param for plots}
\usage{
set_text_size(out, text_size = NULL)
}
\arguments{
\item{out}{config}

\item{text_size}{size of text, will be interperted as pixel}
}
\description{
Get text size param for plots
}
