% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{sort_character}
\alias{sort_character}
\alias{sort_numeric}
\title{Sort a character vector}
\usage{
sort_character(x)

sort_numeric(x)
}
\arguments{
\item{x}{vector of values}
}
\value{
sorted vector
}
\description{
For the numeric sort, all numbers in each string is extracted and we then
sort by the mean of these numbers
}
\section{Functions}{
\itemize{
\item \code{sort_numeric()}: sort a numeric vector

}}
