% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_select}
\alias{theme_select}
\title{Helper to set theme}
\usage{
theme_select(theme_name = getOption("theme"), ...)
}
\arguments{
\item{theme_name}{name of theme}

\item{...}{arguments passed to selected theme}
}
\value{
theme-function to use
}
\description{
Helper to set theme
}
\examples{
theme_select("rc")
}
