% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plots.R
\name{view_highcharts}
\alias{view_highcharts}
\title{Display highcharts object}
\usage{
view_highcharts(plt = NULL)
}
\arguments{
\item{plt}{highcharts object created by rcplot}
}
\description{
Display a highcharts object created på rcplot.
}
