% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plots.R
\name{view_yearly_report_hc}
\alias{view_yearly_report_hc}
\title{Display a highcharts plot from a yearly report created by rcconfig}
\usage{
view_yearly_report_hc(yearly_report = NULL, title = NULL, id = NULL)
}
\arguments{
\item{yearly_report}{A yearly report config created by rcconfig}

\item{title}{Title of graph to display}

\item{id}{id for graph to display}
}
\description{
Display a highcharts plot from a yearly report created by rcconfig
}
