% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{add_total_label}
\alias{add_total_label}
\title{Add total to x-axis}
\usage{
add_total_label(df, total_var = "total", x_var = NULL, break_total = FALSE)
}
\arguments{
\item{df}{data}

\item{total_var}{variable that contains the total}

\item{x_var}{the variable in the x-axis}

\item{break_total}{if there should be a break for the total}
}
\description{
Add total to x-axis
}
