% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{areaspline_highcharts}
\alias{areaspline_highcharts}
\title{Highchart Area Spline plot}
\usage{
areaspline_highcharts(
  df,
  x_var,
  y_var,
  color_var = NULL,
  title = "",
  y_lim = NULL,
  y_breaks = NULL,
  proportion = FALSE,
  scale_percentage = TRUE,
  other_vars = NULL,
  x_lab = NULL,
  y_lab = NULL,
  color_var_order = NULL,
  stacking = NULL,
  fill_opacity = 0.5,
  line_size = 8,
  group_color = NULL,
  legend_title = NULL
)
}
\arguments{
\item{df}{data}

\item{x_var}{x-axis variable}

\item{y_var}{y_axis variable}

\item{color_var}{the name of the color variable}

\item{title}{title of the graph}

\item{y_lim}{limits of y-axis, if \code{NULL} and \code{proportion} is \code{TRUE}, will be
set to \code{c(0, 100)}}

\item{y_breaks}{breaks of y_axis}

\item{proportion}{if the values are proportions or percentages}

\item{scale_percentage}{if proportions should be re-scaled to be percentages}

\item{other_vars}{other variables to include in the tooltip, should be a
named list where the name will be the key in the tooltip}

\item{x_lab}{labels on x axis}

\item{y_lab}{labels on y axis}

\item{color_var_order}{what order \code{color_var} should be displayed in, can
alternatively be \code{auto_character} or \code{auto_numeric} to automatically sort
the levels}

\item{stacking}{stacking mode, one of \code{NULL} (no stacking), \code{"normal"},
or \code{"percent"}}

\item{fill_opacity}{opacity of the area fill, value between 0 and 1}

\item{line_size}{size of the spline border}

\item{group_color}{color of color vars}

\item{legend_title}{title of the legend}
}
\value{
highcharts config
}
\description{
creates an area spline plot in highchart, outputs config for highchart
}
