% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{bar_plot_highcharts}
\alias{bar_plot_highcharts}
\title{Highchart Bar plot}
\usage{
bar_plot_highcharts(
  df,
  x_var,
  y_var,
  horizontal = FALSE,
  fill_var = NULL,
  position = "dodge",
  title = "",
  y_lim = NULL,
  y_breaks = NULL,
  proportion = FALSE,
  scale_percentage = TRUE,
  other_vars = NULL,
  x_lab = NULL,
  y_lab = NULL,
  arrange_by = NULL,
  arrange_desc = TRUE,
  arrange_by_fill = NULL,
  fill_var_order = NULL,
  color_x_value = NULL
)
}
\arguments{
\item{df}{data}

\item{x_var}{x-axis variable}

\item{y_var}{y_axis variable}

\item{horizontal}{whether or not the bar plot should be horizontal,}

\item{fill_var}{the name of the fill variable}

\item{position}{whether or not to make the bars stack or dodge according to
ggplot2 semantics}

\item{title}{title of the graph}

\item{y_lim}{limits of y-axis}

\item{y_breaks}{breaks of y_axis}

\item{proportion}{if the values are percentages}

\item{scale_percentage}{if percentages should the scaled}

\item{other_vars}{other variables to include in the tooltip, should be a
named list where the name will be the key in the tooltip}

\item{x_lab}{labels on x axis}

\item{y_lab}{labels on y axis}
}
\value{
highcharts config
}
\description{
creates a bar plot in highchart, outputs the config file
}
