% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{box_plot_highcharts}
\alias{box_plot_highcharts}
\title{Highchart Box plot}
\usage{
box_plot_highcharts(
  df,
  x_var,
  y_var,
  y_lower,
  y_upper,
  y_min,
  y_max,
  horizontal = FALSE,
  fill_var = NULL,
  position = "dodge",
  title = "",
  y_lim = NULL,
  y_breaks = NULL,
  x_breaks = NULL,
  other_vars = NULL,
  x_lab = NULL,
  y_lab = NULL,
  legend_title = NULL,
  facet_by = NULL,
  group_color = NULL,
  plot_height = NULL,
  bar_size = NULL,
  group_padding = NULL,
  add_total = FALSE,
  total_var = "total",
  multi_total_choice = NULL,
  text_size = NULL
)
}
\arguments{
\item{df}{data}

\item{x_var}{x-axis variable}

\item{y_var}{y_axis variable}

\item{y_lower}{the 25\% value}

\item{y_upper}{the 75\% value}

\item{y_min}{the minimum value}

\item{y_max}{the maximum value}

\item{horizontal}{whether or not the bar plot should be horizontal,}

\item{fill_var}{the name of the fill variable}

\item{position}{whether or not to make the bars stack or dodge according to
ggplot2 semantics}

\item{title}{title of the graph}

\item{y_lim}{limits of y-axis}

\item{y_breaks}{breaks of y_axis}

\item{x_breaks}{breaks of x-axis}

\item{other_vars}{other variables to include in the tooltip, should be a
named list where the name will be the key in the tooltip}

\item{x_lab}{labels on x axis}

\item{y_lab}{labels on y axis}

\item{legend_title}{title of the legend}

\item{facet_by}{variable in \code{df} with at most 2 unique values to facet by;
if supplied the return value is a named list of plots, one per facet level}

\item{group_color}{optional colors}

\item{plot_height}{height of plot, value is in percentages}

\item{bar_size}{width of bars}

\item{group_padding}{distance between bars when using a dodge bar}

\item{add_total}{if total should be added to x-axis}

\item{total_var}{name of column that contains total}

\item{multi_total_choice}{if multiple total values exist for each value in
x_var what value should be picked. choices are min, max, and sum.}

\item{text_size}{size of text, will be interperted as pixels}
}
\value{
highcharts config, or a named list of configs when \code{facet_by} is set
}
\description{
creates highchart box plot, outputs highchart config
}
