% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{box_plot_highcharts}
\alias{box_plot_highcharts}
\title{Highchart Box plot}
\usage{
box_plot_highcharts(
  df,
  x_var,
  y_var,
  y_lower,
  y_upper,
  y_min,
  y_max,
  horizontal = FALSE,
  fill_var = NULL,
  position = "dodge",
  title = "",
  y_lim = NULL,
  y_breaks = NULL,
  other_vars = NULL,
  x_lab = NULL,
  y_lab = NULL
)
}
\arguments{
\item{df}{data}

\item{x_var}{x-axis variable}

\item{y_var}{y_axis variable}

\item{y_lower}{the 25\% value}

\item{y_upper}{the 75\% value}

\item{y_min}{the minimum value}

\item{y_max}{the maximum value}

\item{horizontal}{whether or not the bar plot should be horizontal,}

\item{fill_var}{the name of the fill variable}

\item{position}{whether or not to make the bars stack or dodge according to
ggplot2 semantics}

\item{title}{title of the graph}

\item{y_lim}{limits of y-axis}

\item{y_breaks}{breaks of y_axis}

\item{other_vars}{other variables to include in the tooltip, should be a
named list where the name will be the key in the tooltip}

\item{x_lab}{labels on x axis}

\item{y_lab}{labels on y axis}
}
\value{
highcharts config
}
\description{
creates highchart box plot, outputs highchart config
}
