% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors_select}
\alias{colors_select}
\title{Helper to get colors from theme}
\usage{
colors_select(n, type = "default", theme_name = getOption("theme"))
}
\arguments{
\item{n}{Number of colors needed}

\item{type}{Color palette to use (Paired/default/Spectral)}

\item{theme_name}{name of theme}
}
\value{
a vector with color codes (HEX)
}
\description{
Helper to get colors from theme
}
\examples{
colors_select(3, "default", "rc")
}
