% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{line_plot_highcharts}
\alias{line_plot_highcharts}
\title{Highchart Line plot}
\usage{
line_plot_highcharts(
  df,
  x_var,
  y_var,
  color_var = NULL,
  title = "",
  y_lim = NULL,
  y_breaks = NULL,
  proportion = FALSE,
  scale_percentage = TRUE,
  other_vars = NULL,
  x_lab = NULL,
  y_lab = NULL,
  color_var_order = NULL,
  line_size = 8,
  legend_title = NULL,
  facet_by = NULL,
  group_color = NULL
)
}
\arguments{
\item{df}{data}

\item{x_var}{x-axis variabel}

\item{y_var}{y_axis variabel}

\item{color_var}{the name of the color variable}

\item{title}{title of the graph}

\item{y_lim}{limits of y-axis, if \code{NULL} and \code{proportion} is \code{TRUE}, will be
set to \code{c(0, 100)}}

\item{y_breaks}{breaks of y_axis}

\item{proportion}{if the values are proportions or percentages}

\item{scale_percentage}{if proportions should be re-scaled to be percentages}

\item{other_vars}{other variables to include in the tooltip, should be a
named list where the name will be the key in the tooltip}

\item{x_lab}{labels on x axis}

\item{y_lab}{labels on y axis}

\item{color_var_order}{what order \code{color_var} should be displayed in, can
alternatively be \code{auto_character} or \code{auto_numeric} to automatically sort
the levels}

\item{line_size}{size of lines}

\item{legend_title}{title of the legend}

\item{facet_by}{variable in \code{df} with at most 2 unique values to facet by;
if supplied the return value is a named list of plots, one per facet level}

\item{group_colors}{optional colors}
}
\value{
highcharts config, or a named list of configs when \code{facet_by} is set
}
\description{
creates a line plot in highchart, outputs config for highchart
}
