% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{line_plot_highcharts}
\alias{line_plot_highcharts}
\title{Highchart Line plot}
\usage{
line_plot_highcharts(
  df,
  x_var,
  y_var,
  color_var = NULL,
  title = "",
  y_lim = NULL,
  y_breaks = NULL,
  proportion = FALSE,
  scale_percentage = TRUE,
  other_vars = NULL,
  x_lab = NULL,
  y_lab = NULL,
  color_var_order = NULL,
  line_size = 8
)
}
\arguments{
\item{df}{data}

\item{x_var}{x-axis variabel}

\item{y_var}{y_axis variabel}

\item{color_var}{the name of the color variable}

\item{title}{title of the graph}

\item{y_lim}{limits of y-axis}

\item{y_breaks}{breaks of y_axis}

\item{proportion}{if the values are percentages}

\item{scale_percentage}{if percentages should the scaled}

\item{other_vars}{other variables to include in the tooltip, should be a
named list where the name will be the key in the tooltip}

\item{x_lab}{labels on x axis}

\item{y_lab}{labels on y axis}

\item{color_var_order}{order of color vars}

\item{line_size}{size of lines}
}
\value{
highcharts config
}
\description{
creates a line plot in highchart, outputs config for highchart
}
