% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot.R
\name{map_plot}
\alias{map_plot}
\title{Plot map using geom_sf}
\usage{
map_plot(
  df,
  fill_var = "id",
  label_var = NULL,
  fill_colors = NULL,
  legend_labels = ggplot2::waiver(),
  label_breaks = ggplot2::waiver(),
  legend_row = NULL,
  legend_col = NULL
)
}
\arguments{
\item{df}{data.frame map data. One column
should be named "geometry" containing map
coordinates for regions to map.}

\item{fill_var}{Variable to use for colouring regions}

\item{label_var}{Variable to use label names}

\item{fill_colors}{Manually set colors}

\item{legend_labels}{Manually set labels, character vector.}

\item{label_breaks}{Manually set legend breaks}

\item{legend_row}{Number of rows in legend}

\item{legend_col}{Number of cols in legend}
}
\value{
ggplot object with map figure
}
\description{
Plot map using geom_sf
}
\examples{
# Use internal coordinate datasets
region_coords <- counties
municip_coords <- municipalities

# Map with labels
data <- region_coords \%>\%
  dplyr::mutate(
    labels = paste0(id, " - ", Name)
  )

map_plot(
  df = data,
  fill_var = "labels",
  label_var = "id",
  fill_colors = c(RColorBrewer::brewer.pal(12, "Set3"),
                  RColorBrewer::brewer.pal(9, "Spectral"))
)

# Add some "interresting" data to coordinate dataset and plot
region_coords[["var"]] <- factor(sample(c("Sämst", "Dålig", "Bra", "Bäst"),
                                 size = 21, replace = TRUE),
                                 ordered = TRUE,
                                 levels = c("Sämst", "Dålig", "Bra", "Bäst")
                                 )

map_plot(
  df = region_coords,
  fill_var = "var",
  fill_colors = rc_colors(n = 4, type = "RYG")
)

# Plot both municipalities and region outlines
ggplot2::ggplot() +
  ggplot2::geom_sf(data = municipalities, ggplot2::aes(fill = as.numeric(id))) +
  ggplot2::geom_sf(data = counties, fill = NA, color = "red") +
  ggplot2::theme_void()

# Plot municipalities of Västra Götalands Län, with and without labels
vgregion <- dplyr::filter(municipalities, RegionID == "14")

vgregion[["var"]] <- sample(c("Bäst", "Bra", "Dålig", "Sämst"),
                            size = nrow(vgregion), replace = TRUE)

map_plot(
  df = vgregion,
  fill_var = "var"
)
map_plot(
  df = vgregion,
  fill_var = "var",
  label_var = "Name"
)


}
