% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swe_data.R
\docType{data}
\name{municipalities}
\alias{municipalities}
\title{Map coordinates for Sweden municipalites}
\format{
A sf dataset with one row per municipality
\describe{
\item{id}{ID of municipality}
\item{geometry}{Coordinate polygon used to draw municipality borders}
}
}
\source{
\url{http://api.thenmap.net/v2/se-7/geo/2020-06-03}
}
\usage{
municipalities
}
\description{
Map coordinates for Sweden municipalites
}
\examples{
library(sf)
library(ggplot2)

ggplot() +
 geom_sf(data = municipalities, aes(fill = as.numeric(id))) +
 theme_void()

ggplot() +
  geom_sf(data = municipalities, aes(fill = as.numeric(id))) +
  geom_sf(data = counties, fill = NA, color = "red") +
 theme_void()

}
\keyword{datasets}
