% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{bar_plot_highcharts}
\alias{bar_plot_highcharts}
\alias{line_plot_highcharts}
\alias{plot_highcharts}
\alias{add_y_axis}
\alias{add_tooltip}
\alias{make_series}
\title{Create a config for a highchart plot}
\usage{
bar_plot_highcharts(
  df,
  x_var,
  y_var,
  horizontal = FALSE,
  fill_var = NULL,
  position = "dodge",
  title = "",
  y_lim = NULL,
  y_breaks = NULL,
  proportion = FALSE,
  scale_percentage = TRUE,
  other_vars = NULL,
  x_lab = NULL,
  y_lab = NULL
)

line_plot_highcharts(
  df,
  x_var,
  y_var,
  color_var = NULL,
  title = "",
  y_lim = NULL,
  y_breaks = NULL,
  proportion = FALSE,
  scale_percentage = TRUE,
  other_vars = NULL,
  x_lab = NULL,
  y_lab = NULL
)

plot_highcharts(
  df,
  x_var,
  vars,
  title,
  group_vars,
  y_lim,
  y_breaks,
  proportion = FALSE,
  scale_percentage = FALSE,
  type,
  other_vars,
  horizontal = FALSE,
  x_lab = NULL,
  y_lab = NULL
)

add_y_axis(
  out,
  y_lim = NULL,
  y_breaks = NULL,
  proportion = FALSE,
  y_lab = NULL
)

add_tooltip(
  out,
  df,
  group_vars = NULL,
  other_vars = NULL,
  proportion = FALSE,
  type
)

make_series(
  df,
  vars,
  group_vars = NULL,
  palette_type = "qualitative",
  colors = NULL,
  other_vars = NULL,
  proportion = FALSE,
  scale_percentage = TRUE
)
}
\arguments{
\item{df}{the data.frame to plot}

\item{x_var}{the name of the x variable'}

\item{y_var}{the name of the y variable}

\item{horizontal}{whether or not the bar plot should be horizontal,}

\item{fill_var}{the name of the fill variable}

\item{position}{whether or not to make the bars stack or dodge according to
ggplot2 semantics}

\item{title}{the title of plot}

\item{y_lim}{limits for the y-axis}

\item{y_breaks}{vector of breaks for y-axis}

\item{proportion}{whether the y variable is a proportion}

\item{scale_percentage}{whether or not to scale the y-variable by 100}

\item{other_vars}{other variables to include in the tooltip, should be a
named list where the name will be the key in the tooltip}

\item{x_lab}{the title of the x axis}

\item{y_lab}{the title of the y axis}

\item{color_var}{the name of the color variable}

\item{vars}{a named list used to rename variables to proper highcharts keys}

\item{group_vars}{the variables indicating a series to plot}

\item{type}{the type of graph}

\item{out}{intermediate config}

\item{palette_type}{passed to \code{\link[=colors_rc_2]{colors_rc_2()}}}

\item{colors}{an optional subset of \code{\link[=colors_rc_2]{colors_rc_2()}} with
\code{palette_type = "qualitative"} and \code{n = 9} used to manually set colors}
}
\value{
highcharts config
}
\description{
Create a config for a highchart plot
}
\section{Functions}{
\itemize{
\item \code{bar_plot_highcharts()}: creates a bar plot

\item \code{line_plot_highcharts()}: creates a line plot

\item \code{add_y_axis()}: add the y axis to the config

\item \code{add_tooltip()}: add the tooltip to the config

\item \code{make_series()}: converts a data.frame to a highcharts series
representation

}}
