% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar_plot.R
\name{radar_plot}
\alias{radar_plot}
\title{Produce a radar plot}
\usage{
radar_plot(
  df,
  seg = NULL,
  colors = rcplot::colors_rc_2(nrow(df)),
  y_max = NULL,
  y_min = NULL
)
}
\arguments{
\item{df}{A data frame containing the values for the radar plot.}

\item{seg}{Number of segments in the radar chart. If \code{NULL}, it is
automatically set based on \code{pretty(y_min:y_max)}.}

\item{colors}{A vector of colors used for the radar chart lines. Defaults to
\code{rcplot::colors_rc_2(nrow(df))}.}

\item{y_max}{Maximum value for the radar chart axes. If \code{NULL}, it is
automatically set to the maximum value in \code{df} plus one.}

\item{y_min}{Minimum value for the radar chart axes. Defaults to \code{0}.}
}
\value{
A radar chart plotted using \code{fmsb::radarchart()}.
}
\description{
Produces a radar plot using \code{fmsb::radarchart()}, with automatic handling
of axis limits and colors.
}
\examples{
df <- data.frame(A = c(3, 4, 5), B = c(2, 3, 4), C = c(4, 5, 3))
radar_plot(df)
}
