% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{rc_colors}
\alias{rc_colors}
\title{Wrapper for RColorBrewer to get colors}
\usage{
rc_colors(n = 3, type = "default")
}
\arguments{
\item{n}{number of colors}

\item{type}{one of
\code{"default"}, for standard colors.
\code{"Spectral"}, scale.
\code{"Paired"} for pairwise colors.}
}
\value{
vector with color codes
}
\description{
Wrapper for RColorBrewer to get colors
}
\examples{
rc_colors(2)
}
