% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rcsave}
\alias{rcsave}
\title{ggsave function with default settings as specified with options()}
\usage{
rcsave(
  file,
  plot = ggplot2::last_plot(),
  width = getOption("figure_width"),
  height = getOption("figure_height"),
  units = "in",
  dpi = getOption("figure_dpi"),
  device = getOption("figure_device"),
  family = getOption("figure_family"),
  ...
)
}
\arguments{
\item{file}{File name to create on disk.}

\item{plot}{Plot to save}

\item{width}{Width in units ("in", "cm", or "mm").
default: options("figure_width")}

\item{height}{Height in units ("in", "cm", or "mm").}

\item{units}{Units ("in", "cm", or "mm")}

\item{dpi}{Plot resolution.}

\item{device}{Device to use. Can either be a device function (e.g. png()), or one of "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{family}{Font family}

\item{...}{additional arguments passed to \code{ggplot2::ggsave} and
then to the selected \code{device}.}
}
\description{
ggsave function with default settings as specified with options()
}
