% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{replace_categories_with_years}
\alias{replace_categories_with_years}
\title{Convert x axis days to years}
\usage{
replace_categories_with_years(plt, type = "years")
}
\arguments{
\item{type}{which type of labels, if 'years' will only display the years.}

\item{out}{config}
}
\value{
highcarts config
}
\description{
This function is used for survival curves to changes the days values in the
x-axis to years.
}
