% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{set_size_params}
\alias{set_size_params}
\title{Create parameters for sizing and padding for bar plot.}
\usage{
set_size_params(
  out,
  position,
  bar_size = NULL,
  plot_height = NULL,
  group_padding = NULL
)
}
\arguments{
\item{out}{config}

\item{position}{if bar should be stacked or dodge}
}
\description{
This function dynamically calculates bar width, chart height, and padding to
improve readability for both horizontal and vertical plots, while also
allowing manual overrides for bar size, plot height, and group padding.
}
\details{
The parameters take into account if it is a stacked or dodge plot and for
data with many categories it tries to adjust the group padding in such a way
that the bars dont overlap.

An important thing is to keep the distance between the bars inside the
x-values smaller than the distance between the grouped bars between the
x_values.
}
